Вынесет компонент рейтинга в отдельный файл
