+++ src/components/comment-form/comment-form.tsx
@@ -2,17 +2,11 @@ import { FormEvent } from 'react';
 import { useAppDispatch, useAppSelector } from '../../hooks/redux-ts';
 import { addReview } from '../../store/api-actions';
 import { changeComment, changeRating } from '../../store/reviews-data/reviews-action';
+import RatingInput from '../rating-input/rating-input';
 
 type RatingInput = {
   value: number;
   title: string;
-}
-
-type RatingInputProps = {
-  rating: number | null;
-  value: number;
-  title: string;
-  handleInputChange: (value: number) => void | undefined;
 }
 
 type CommentFormProps = {
@@ -44,27 +38,6 @@ const ratingInputs: RatingInput[] = [
 
 const INITIAL_RATING = null;
 const MIN_COMMENT_LENGTH = 50;
-
-function RatingInput({ rating, value, title, handleInputChange }: RatingInputProps): React.JSX.Element {
-  return (
-    <>
-      <input
-        className="form__rating-input visually-hidden"
-        name="rating"
-        value={value}
-        id={`${value}-stars`}
-        type="radio"
-        onChange={() => handleInputChange(Number(value))}
-        checked={value === rating}
-      />
-      <label htmlFor={`${value}-stars`} className="reviews__rating-label form__rating-label" title={title}>
-        <svg className="form__star-image" width="37" height="33">
-          <use xlinkHref="#icon-star"></use>
-        </svg>
-      </label>
-    </>
-  );
-}
 
 export default function CommentForm({ offerId }: CommentFormProps): React.JSX.Element {
   const dispatch = useAppDispatch();
